pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "readwise-wiki"
version = "0.1.0"
description = "Ingest Readwise highlights and emit a Wikiwise-compatible LLM wiki"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.1,<6.0",
  "celery[redis]>=5.4",
  "redis>=5.0",
  "psycopg[binary]>=3.2",
  "django-environ>=0.11",
  "requests>=2.32",
  "openai>=1.50",
  "rapidfuzz>=3.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-django>=4.9",
  "pytest-cov>=5.0",
  "factory-boy>=3.3",
  "responses>=0.25",
  "ruff>=0.7",
  "pre-commit>=4.0",
]

[tool.setuptools.packages.find]
include = ["config*", "wiki*"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
  "E", "F", "W",   # pycodestyle / pyflakes
  "I",             # isort
  "B",             # bugbear
  "UP",            # pyupgrade
  "SIM",           # simplify
  "DJ",            # django
]
ignore = [
  "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "DJ"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--cov=wiki",
  "--cov-report=term-missing",
]
filterwarnings = [
  "ignore::DeprecationWarning:celery.*",
]

[tool.coverage.run]
source = ["wiki"]
omit = [
  "*/migrations/*",
  "*/tests/*",
  "wiki/management/__init__.py",
]
